refactor: Update Searchbar component to include a clear icon for CPF field
The Searchbar component in the Dashboard has been updated to include a clear icon for the CPF field. This allows users to easily reset the CPF input and trigger a query invalidation for registration data. The `handleResetCpfField` function has been added to reset the field and call the `onInvalidateRegistrationQuery` function. This improves the user experience by providing a convenient way to clear the input and start a new search.