fix: Update getAdmissionDate to include time in admission date
The getAdmissionDate function in the src/repositories/helpers/get-admission-date.ts file has been updated to include the time in the admission date. This change ensures that the formatted admission date includes the time as well, by appending 'T00:00:00' to the input date.

Refactor the getAdmissionDate function to include time in the admission date